+++ src/Resources/app/administration/src/component/nosto-job-sub-jobs/index.js
@@ -2,142 +2,7 @@
  * @sw-package innovation
  */
 
-import template from './nosto-job-sub-jobs.html.twig';
-import JobHelper from '../../util/job.helper';
-import './nosto-job-sub-jobs.scss';
-
 const { Component } = Shopware;
-const { Criteria } = Shopware.Data;
 
 /** @private */
-Component.register('nosto-job-sub-jobs', {
-    template,
-
-    inject: [
-        'NostoRescheduleService',
-        'repositoryFactory',
-    ],
-
-    mixins: [
-        'notification',
-    ],
-
-    props: {
-        jobId: {
-            type: String,
-            required: false,
-            default: null,
-        },
-    },
-
-    data() {
-        return {
-            subJobs: null,
-            showMessagesModal: false,
-            currentJobMessages: null,
-            page: 1,
-            limit: 25,
-        };
-    },
-
-    computed: {
-        jobRepository() {
-            return this.repositoryFactory.create('nosto_scheduler_job');
-        },
-
-        jobChildrenColumns() {
-            return [
-                {
-                    property: 'name',
-                    dataIndex: 'name',
-                    label: this.$tc('job-listing.page.listing.grid.column.name'),
-                    allowResize: false,
-                    inlineEdit: true,
-                    width: '200px',
-                },
-                {
-                    property: 'status',
-                    dataIndex: 'status',
-                    label: this.$tc('job-listing.page.listing.grid.column.status'),
-                    allowResize: false,
-                    inlineEdit: true,
-                    width: '100px',
-                },
-                {
-                    property: 'startedAt',
-                    dataIndex: 'startedAt',
-                    label: this.$tc('job-listing.page.listing.grid.column.started-at'),
-                    allowResize: false,
-                    inlineEdit: true,
-                    width: '150px',
-                    sortable: true,
-                },
-                {
-                    property: 'finishedAt',
-                    dataIndex: 'finishedAt',
-                    label: this.$tc('job-listing.page.listing.grid.column.finished-at'),
-                    allowResize: true,
-                    inlineEdit: true,
-                    width: '150px',
-                },
-                {
-                    property: 'createdAt',
-                    dataIndex: 'createdAt',
-                    label: this.$tc('job-listing.page.listing.grid.column.created-at'),
-                    allowResize: true,
-                    inlineEdit: true,
-                    width: '150px',
-                },
-                {
-                    property: 'messages',
-                    dataIndex: 'messages',
-                    label: 'Messages',
-                    allowResize: true,
-                    inlineEdit: false,
-                    width: '250px',
-                    sortable: false,
-                },
-            ];
-        },
-    },
-
-    created() {
-        this.initModalData();
-    },
-
-    methods: {
-        initModalData() {
-            const criteria = new Criteria(this.page, this.limit);
-            criteria.addFilter(Criteria.equals('parentId', this.jobId));
-            criteria.addSorting(Criteria.sort('createdAt', 'DESC', false));
-            criteria.addAssociation('messages');
-            this.jobRepository.search(criteria, Shopware.Context.api).then(jobItems => {
-                this.subJobs = JobHelper.sortMessages(jobItems);
-            });
-        },
-
-        rescheduleJob(jobId) {
-            this.NostoRescheduleService.rescheduleJob(jobId).then(() => {
-                this.createNotificationSuccess({
-                    message: 'Job has been rescheduled successfully.',
-                });
-                this.initPageData();
-            }).catch(() => {
-                this.createNotificationError({
-                    message: 'Unable reschedule job.',
-                });
-            });
-        },
-
-        showMessageModal(messages) {
-            this.currentJobMessages = messages;
-            this.showMessagesModal = true;
-        },
-
-        getMessagesCount(job, type) {
-            return job.messages.filter((item) => {
-                return item.type === `${type}-message`;
-            }).length;
-        },
-    },
-});
+Component.register('nosto-job-sub-jobs', () => import('./component'));
